Modified to check multiple images and check with original labels
Added the posibility of processing multiple images and comparing against the same images processed as in label_images.py. The idea is to check if the SSD is good against the labels or not. Another question is if the labels are OK. So the idea is to made two checks to be able to argue our procedure :

- Our label algorithm is quite ok
- The SSD replicates the labeling algorithm

Also it is useful to give Frank the most interesting images to relabel manually.